+++ src/map/get_at.c
@@ -1,25 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   [TEMP]map_get_at.c                                 :+:    :+:            */
+/*   get_at.c                                           :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: rdrazsky <rdrazsky@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/02/22 13:37:12 by rdrazsky      #+#    #+#                 */
-/*   Updated: 2022/02/22 13:37:14 by rdrazsky      ########   odam.nl         */
+/*   Updated: 2022/02/24 18:46:54 by rdrazsky      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <cub3d.h>
 
-/* TO BE DELETED */
 char	map_get_at(t_ft_list *map, UINT x, UINT y)
 {
 	t_ft_string	*line;
 
 	if (x < 0 || y < 0)
 		return ('\0');
-	if (y >= map->len)
+	if (y >= (UINT)map->len)
 		return ('\0');
 	line = ft_list_item_at(map, y);
 	if (x >= line->len)
